halaman admin

+++ admin/berita.php
@@ -1,4 +1,6 @@
 <?php
+require_once "../helpers/helpers.php";
+
 ob_start();
 $hostname       = "localhost";
 $user           = "root";
@@ -9,9 +11,10 @@ $judul          = "";
 $deskripsi      = "";
 $lokasi         = "";
 $tanggal        = "";
-$gambar          = "";
+$gambar         = "";
 $success        = "";
 $failed         = "";
+$op             = "";
 
 $koneksi = mysqli_connect($hostname, $user, $pwd, $db);
 if (!$koneksi) {
@@ -26,8 +29,22 @@ if (isset($_GET['op'])) {
     $op = "";
 }
 
-if ($op == 'hapus') {
-    $id         = $_GET['id'];
+if ($op == 'edit' && isset($_GET['id']) && !empty($_GET['id'])) {
+    $id = $_GET['id'];
+    $id         = enkripsiUrl('decrypt', $id);
+    $sql1       = "SELECT * FROM tb_berita WHERE id = '$id'";
+    $q1         = mysqli_query($koneksi, $sql1);
+    $r1         = mysqli_fetch_array($q1);
+    $judul      = $r1['judul'];
+    $deskripsi  = $r1['deskripsi'];
+    $lokasi     = $r1['lokasi'];
+    $tanggal    = $r1['tanggal'];
+    $gambar     = $r1['gambar'];
+}
+
+if ($op == 'hapus' && isset($_GET['id']) && !empty($_GET['id'])) {
+    $id = $_GET['id'];
+    $id         = enkripsiUrl('decrypt', $id);
     $sql1       = "DELETE FROM tb_berita WHERE id = '$id'";
     $q1         = mysqli_query($koneksi, $sql1);
     if ($q1) {
@@ -35,18 +52,6 @@ if ($op == 'hapus') {
     } else {
         $failed  = "Gagal menghapus data berita";
     }
-}
-
-if ($op == 'edit') {
-    $id         = $_GET['id'];
-    $sql1       = "SELECT * FROM tb_berita WHERE id = '$id'";
-    $q1         = mysqli_query($koneksi, $sql1);
-    $r1         = mysqli_fetch_array($q1);
-    $judul      = $r1['judul'];
-    $deskripsi  = $r1['deskripsi'];
-    $lokasi     = $r1['lokasi'];
-    $tanggal    = $r1['tanggal'];   
-    $gambar     = $r1['gambar'];
 }
 
 if (isset($_POST['simpan'])) { //untuk create data
@@ -61,7 +66,7 @@ if (isset($_POST['simpan'])) { //untuk create data
     move_uploaded_file($tmp_name, $path . $gambar);
 
     if ($judul && $deskripsi && $lokasi && $tanggal && $gambar) {
-        if ($op == 'edit') { //untuk update
+        if ($op == 'edit' && isset($_GET['id'])) { //untuk update
             $sql1       = "UPDATE tb_berita SET judul = '$judul', deskripsi = '$deskripsi', lokasi = '$lokasi', tanggal = '$tanggal', gambar = '$gambar'";
             $sql1       .= "WHERE id = '$id'";
             $q1         = mysqli_query($koneksi, $sql1);
@@ -125,12 +130,12 @@ if (isset($_POST['simpan'])) { //untuk create data
                 <div class="menu-inner">
                     <nav>
                         <ul class="metismenu" id="menu">
-                            <li><a href="./index.php" aria-expanded="true"><i class="ti-dashboard"></i><span>Dashboard</span></a></li>
-                            <li><a href="./berita.php" aria-expanded="true"><i class="ti-world"></i><span>Berita</span></a></li>
-                            <li><a href="./mitra.php" aria-expanded="true"><i class="ti-user"></i><span>Mitra</span></a></li>
-                            <li><a href="./kerjasama.php" aria-expanded="true"><i class="ti-link"></i><span>Kerjasama</span></a></li>
-                            <li><a href="./galeri.php" aria-expanded="true"><i class="ti-announcement"></i><span>Galeri</span></a></li>
-                            <li><a href="./tentang.php" aria-expanded="true"><i class="ti-headphone"></i><span>Tentang</span></a></li>
+                            <li><a href="../admin/" aria-expanded="true"><i class="ti-dashboard"></i><span>Dashboard</span></a></li>
+                            <li><a href="./berita" aria-expanded="true"><i class="ti-world"></i><span>Berita</span></a></li>
+                            <li><a href="./mitra" aria-expanded="true"><i class="ti-user"></i><span>Mitra</span></a></li>
+                            <li><a href="./kerjasama" aria-expanded="true"><i class="ti-link"></i><span>Kerjasama</span></a></li>
+                            <li><a href="./galeri" aria-expanded="true"><i class="ti-announcement"></i><span>Galeri</span></a></li>
+                            <li><a href="./tentang" aria-expanded="true"><i class="ti-headphone"></i><span>Tentang</span></a></li>
                         </ul>
                     </nav>
                 </div>
@@ -141,6 +146,11 @@ if (isset($_POST['simpan'])) { //untuk create data
                 <div class="row align-items-center">
                     <div class="col-sm-6">
                         <div class="breadcrumbs-area clearfix">
+                            <div class="nav-btn pull-left">
+                                <span></span>
+                                <span></span>
+                                <span></span>
+                            </div>
                             <h4 class="page-title pull-left">Berita</h4>
                             <ul class="breadcrumbs pull-left">
                                 <li><a href="index.html">Admin</a></li>
@@ -149,10 +159,10 @@ if (isset($_POST['simpan'])) { //untuk create data
                         </div>
                     </div>
                     <div class="col-sm-6 clearfix">
-                        <div class="user-progambar pull-right">
+                        <div class="user-profile pull-right">
                             <h4 class="user-name dropdown-toggle" data-toggle="dropdown">Username Admin<i class="fa fa-angle-down"></i></h4>
                             <div class="dropdown-menu">
-                                <a class="dropdown-item" href="./auth.php">Log Out</a>
+                                <a class="dropdown-item" href="./auth">Log Out</a>
                             </div>
                         </div>
                     </div>
@@ -168,14 +178,14 @@ if (isset($_POST['simpan'])) { //untuk create data
                         <?php echo $success ?>
                     </div>
                 <?php
-                    header("refresh:3; url=berita.php");
+                    header("refresh:3; url=berita");
                 } elseif ($failed) {
                 ?>
                     <div class="alert alert-danger" role="alert">
                         <?php echo $failed ?>
                     </div>
                 <?php
-                    header("refresh:3; url=berita.php");
+                    header("refresh:3; url=berita");
                 }
                 ?>
                 <form method="post" enctype="multipart/form-data">
@@ -215,7 +225,7 @@ if (isset($_POST['simpan'])) { //untuk create data
                             <input type="file" name="gambar" id="gambar" class="form-control" required value="<?php echo $gambar ?>">
                         </div>
                     </div>
-                    <input type="submit" value="Simpan" name="simpan" class="btn btn-primary text-uppercase col-2 offset-2">
+                    <input type="submit" value="Simpan" name="simpan" class="btn btn-primary text-uppercase col-lg-10 col-12 offset-lg-2">
                 </form>
             </div>
             <section id="record" class="py-5 my-5">
@@ -236,7 +246,7 @@ if (isset($_POST['simpan'])) { //untuk create data
                             $sql   = "select * from tb_berita order by id";
                             $query    = mysqli_query($koneksi, $sql);
                             while ($q = mysqli_fetch_array($query)) {
-                                $id         = $q['id'];
+                                $id         = enkripsiUrl('encrypt', $q['id']);
                                 $judul      = $q['judul'];
                                 $deskripsi  = $q['deskripsi'];
                                 $lokasi     = $q['lokasi'];
@@ -247,16 +257,16 @@ if (isset($_POST['simpan'])) { //untuk create data
                                     <td scope="row"><?php echo $judul ?></td>
                                     <td scope="row"><?php echo $deskripsi ?></td>
                                     <td scope="row"><?php echo $lokasi ?></td>
-                                    <td scope="row"><?php echo $tanggal ?></td>
+                                    <td scope="row"><?php echo date('d F Y', strtotime($tanggal)) ?></td>
                                     <td scope="row">
                                         <img src="../admin/assets/upload/berita/<?php echo $gambar ?>" style="width: 150px;" class="img-fluid w-0">
                                     </td>
                                     <td class="text-center">
-                                        <a href="berita.php?op=edit&id=<?php echo $id ?>">
-                                            <button type="button" name="edit" class="col-4 btn btn-warning">Edit</button>
+                                        <a href="berita?op=edit&id=<?php echo $id ?>">
+                                            <button type="button" name="edit" class="col-lg-4 col-10 btn btn-warning">Edit</button>
                                         </a>
-                                        <a href="berita.php?op=hapus&id=<?php echo $id ?>" onclick="return confirm('Apakah yakin mau delete data?')">
-                                            <button type="button" name="hapus" class="col-4 btn btn-danger">Delete</button>
+                                        <a href="berita?op=hapus&id=<?php echo $id ?>" onclick="return confirm('Apakah yakin mau delete data?')">
+                                            <button type="button" name="hapus" class="col-lg-4 col-10 btn btn-danger">Delete</button>
                                         </a>
                                     </td>
                                 </tr>
